venv5/requirements.txt
apache-beam[gcp]==2.40.0
mysql-connector-python==8.0.31
lxml==4.9.2